+++ core/web/resolver/node_test.go
@@ -5,13 +5,104 @@ import (
 	"encoding/json"
 	"testing"
 
+	gqlerrors "github.com/graph-gophers/graphql-go/errors"
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/guregu/null.v4"
 
 	"github.com/smartcontractkit/chainlink/core/chains/evm"
 	"github.com/smartcontractkit/chainlink/core/chains/evm/types"
 	"github.com/smartcontractkit/chainlink/core/utils"
+	"github.com/smartcontractkit/chainlink/core/utils/stringutils"
 )
+
+func TestResolver_Nodes(t *testing.T) {
+	t.Parallel()
+
+	var (
+		chainID = *utils.NewBigI(1)
+		nodeID  = int32(200)
+
+		query = `
+			query GetNodes {
+				nodes {
+					results {
+						id
+						name
+						createdAt
+						chain {
+							id
+						}
+					}
+					metadata {
+						total
+					}
+				}
+			}`
+	)
+	gError := errors.New("error")
+
+	testCases := []GQLTestCase{
+		unauthorizedTestCase(GQLTestCase{query: query}, "nodes"),
+		{
+			name:          "success",
+			authenticated: true,
+			before: func(f *gqlTestFramework) {
+				f.Mocks.evmORM.On("Nodes", PageDefaultOffset, PageDefaultLimit).Return([]types.Node{
+					{
+						ID:         nodeID,
+						Name:       "node-name",
+						EVMChainID: chainID,
+						CreatedAt:  f.Timestamp(),
+					},
+				}, 1, nil)
+				f.Mocks.evmORM.On("GetChainsByIDs", []utils.Big{chainID}).Return([]types.Chain{
+					{
+						ID: chainID,
+					},
+				}, nil)
+				f.App.On("EVMORM").Return(f.Mocks.evmORM)
+			},
+			query: query,
+			result: `
+			{
+				"nodes": {
+					"results": [{
+						"id": "200",
+						"name": "node-name",
+						"createdAt": "2021-01-01T00:00:00Z",
+						"chain": {
+							"id": "1"
+						}
+					}],
+					"metadata": {
+						"total": 1
+					}
+				}
+			}`,
+		},
+		{
+			name:          "generic error",
+			authenticated: true,
+			before: func(f *gqlTestFramework) {
+				f.Mocks.evmORM.On("Nodes", PageDefaultOffset, PageDefaultLimit).Return([]types.Node{}, 0, gError)
+				f.App.On("EVMORM").Return(f.Mocks.evmORM)
+			},
+			query:  query,
+			result: `null`,
+			errors: []*gqlerrors.QueryError{
+				{
+					Extensions:    nil,
+					ResolverError: gError,
+					Path:          []interface{}{"nodes"},
+					Message:       gError.Error(),
+				},
+			},
+		},
+	}
+
+	RunGQLTests(t, testCases)
+}
 
 func Test_NodeQuery(t *testing.T) {
 	t.Parallel()
@@ -186,7 +277,7 @@ func Test_DeleteNodeMutation(t *testing.T) {
 	}
 
 	variables := map[string]interface{}{
-		"id": fakeID,
+		"id": stringutils.FromInt32(fakeID),
 	}
 
 	d, err := json.Marshal(map[string]interface{}{