+++ operator_ui/src/Private.tsx
@@ -10,10 +10,7 @@ import Notifications from 'pages/Notifications'
 import PrivateRoute from './PrivateRoute'
 
 import DashboardIndex from 'pages/Dashboards/Index'
-import BridgesNew from 'pages/Bridges/New'
-import BridgesEdit from 'pages/Bridges/Edit'
 import Configuration from 'pages/Configuration/Index'
-import JobsIndex from 'pages/JobsIndex/JobsIndex'
 import JobsShow from 'pages/Jobs/Show'
 import JobsNew from 'pages/Jobs/New'
 import JobRunsIndex from 'pages/JobRuns/Index'
@@ -21,17 +18,17 @@ import JobRunsShowOverview from 'pages/Jobs/Runs/Show'
 import ChainsIndex from 'pages/ChainsIndex/ChainsIndex'
 import ChainsNew from 'pages/Chains/New'
 import ChainShow from 'pages/Chains/Show'
-import NodeShow from 'pages/Nodes/Show'
+import { NodeScreen } from 'screens/Node/NodeScreen'
 import KeysIndex from 'pages/Keys/Index'
 import NotFound from 'pages/NotFound'
 import TransactionsIndex from 'pages/Transactions/Index'
 import TransactionsShow from 'pages/Transactions/Show'
 import NodesIndex from './pages/NodesIndex/NodesIndex'
 
-import { BridgesPage } from 'pages/Bridges' // Rename to lower case when completed
-
+import { BridgesPage } from 'pages/bridges'
+import { JobsPage } from 'pages/JobsIndex'
 import { FeedsManagerPage } from 'pages/feeds_manager'
-import { JobProposalsPage } from './pages/job_proposals'
+import { JobProposalsPage } from 'pages/job_proposals'
 
 const styles = (theme: Theme) => {
   return {
@@ -70,9 +67,6 @@ const Private = ({ classes }: { classes: { content: string } }) => {
                 />
               </PrivateRoute>
 
-              <PrivateRoute exact path="/jobs">
-                <JobsIndex />
-              </PrivateRoute>
               <PrivateRoute exact path="/jobs/new">
                 <JobsNew />
               </PrivateRoute>
@@ -116,17 +110,10 @@ const Private = ({ classes }: { classes: { content: string } }) => {
                 <NodesIndex />
               </PrivateRoute>
 
-              <PrivateRoute path="/nodes/:nodeId">
-                <NodeShow />
+              <PrivateRoute path="/nodes/:id">
+                <NodeScreen />
               </PrivateRoute>
 
-              <PrivateRoute exact path="/bridges/new" component={BridgesNew} />
-
-              <PrivateRoute
-                exact
-                path="/bridges/:bridgeId/edit"
-                component={BridgesEdit}
-              />
               <PrivateRoute
                 exact
                 path="/transactions"
@@ -155,6 +142,10 @@ const Private = ({ classes }: { classes: { content: string } }) => {
 
               <PrivateRoute path="/job_proposals">
                 <JobProposalsPage />
+              </PrivateRoute>
+
+              <PrivateRoute exact path="/jobs">
+                <JobsPage />
               </PrivateRoute>
 
               <PrivateRoute component={NotFound} />