+++ operator_ui/src/screens/Bridges/BridgesScreen.tsx
@@ -33,7 +33,7 @@ export const BridgesScreen: React.FC = () => {
     FetchBridgesVariables
   >(BRIDGES_QUERY, {
     variables: { offset: (page - 1) * pageSize, limit: pageSize },
-    nextFetchPolicy: 'cache-and-network',
+    fetchPolicy: 'network-only',
   })
 
   if (loading) {